Make UI sense and adjust ideaView pipe line
