+++ miniprogram/app.wxss
@@ -1,5 +1,6 @@
 /**app.wxss**/
 @import "style/weui.wxss";
+@import "style/theme.wxss";
 .container {
   display: flex;
   flex-direction: column;
