+++ miniprogram/class/Idea.js
@@ -1,5 +1,3 @@
-import { CLOUD_FILE_HEAD, MAX_FETCH_URL_COUNT } from './Constants'
-import { MediaType } from './IdeaType'
 const app = getApp()
 
 class Idea {
@@ -82,61 +80,6 @@ class Idea {
       title: this.title,
       iconPath: iconPath
     })
-  }
-
-  /**
-   * 将想法子项中的 cloudID 替换成 tempUrl
-   * @param {*} items 想法子项
-   * @returns itemId2SwapSrc itemId 到 tempUrl 的映射
-   */
-  static async replaceCloudID2TempUrl (items) {
-    // 初始化非 markdown 类想法子项 idea 备份
-    // 获取 coudId 到 uuid _d 的映射
-    const cloudId2uuid = new Map()
-    for (const i of items) {
-      if (i.src.startsWith(CLOUD_FILE_HEAD)) {
-        cloudId2uuid.set(i.src, i._id)
-      }
-    }
-    // 获取分割的 cloudID list
-    // getTempUrl 每次最多获取 MAX_FETCH_URL_COUNT 个文件 url
-    const MARKDOWN = (new MediaType()).MARKDOWN
-    const tmpList = items.filter(item => item.type !== MARKDOWN)
-    const fetchList = []
-    const oneTimeFetchList = []
-    let cnt = 0
-    for (const i in tmpList) {
-      if (cnt === MAX_FETCH_URL_COUNT) {
-        fetchList.push(oneTimeFetchList)
-        oneTimeFetchList.length = 0
-        cnt = 0
-      }
-      oneTimeFetchList.push(tmpList[i].src)
-    }
-    if (oneTimeFetchList.length !== 0) {
-      fetchList.push(oneTimeFetchList)
-    }
-    // 设置 uuid 到 换取的src 的映射
-    const itemId2SwapSrc = new Map()
-    for (const i of fetchList) {
-      try {
-        const res = await wx.cloud.getTempFileURL({
-          fileList: i
-        })
-        for (const j of res.fileList) {
-          itemId2SwapSrc.set(cloudId2uuid.get(j.fileID), j.tempFileURL)
-        }
-      } catch (err) {
-        console.log(err)
-      }
-    }
-    // 将 换取的src 设置到 items 上
-    for (const i of items) {
-      if (itemId2SwapSrc.has(i._id)) {
-        i.src = itemId2SwapSrc.get(i._id)
-      }
-    }
-    return itemId2SwapSrc
   }
 }
 export {
